requirements.txt
python-telegram-bot==20.7
pymongo==4.6.0
python-dotenv==1.0.0
Pillow==10.1.0
pytesseract==0.3.10
opencv-python==4.8.1.78
numpy==1.24.3
easyocr==1.7.0
scikit-image==0.21.0
aiofiles==23.2.1